+++ src/pages/CartNLikesPage/index.tsx
@@ -1,14 +1,15 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import logo from '../../assets/008.png';
-import vector from '../../assets/Vector.svg';
 import CartList from '../../components/CartNLikePage/CarList';
 import { Navigate } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { autoCheck } from '../../features/authSlice';
+import useUserInfo from '../../hooks/useUserInfo';
 type Props = {};
 
 const CartNLikesPage = (props: Props) => {
   const isLogin = useSelector((state: autoCheck) => state.auth.isAuthenticated);
+  const user = useUserInfo();
 
   return isLogin ? (
     <div className='relative'>
@@ -18,7 +19,7 @@ const CartNLikesPage = (props: Props) => {
         className='opacity-[0.2] absolute top-0 right-0 w-[227px] h-[191px]'
       />
       <div className='mt-5 text-center text-[30px] font-bold'>
-        <span className='text-mw'>미왕이</span>님 상품몰 장바구니
+        <span className='text-mw'>{user?.name}</span>님 상품몰 장바구니
       </div>
       <CartList />
     </div>
